requirements.txt
Flask==2.0.2
Flask-APScheduler==1.12.4
Flask-Migrate==2.1.1
Flask-Login==0.4.0
Flask_Reuploaded==1.2.0
Flask-SQLAlchemy==2.5.1
flask-marshmallow==0.14.0
marshmallow-sqlalchemy==0.27.0
PyMySQL==1.0.2
psutil==5.9.4
Flask-Mail==0.9.1
sqlparse==0.2.3
Pillow==8.4.0
python-dotenv==0.19.1
validators==0.20.0